requirements.txt
APScheduler==3.6.1
cachetools==3.1.1
certifi==2019.9.11
Click==7.0
Flask==1.1.1
Flask-HTTPAuth==3.3.0
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
gspread==3.1.0
gunicorn==20.0.0
httplib2>=0.18.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
oauth2client==4.1.3
oauthlib==3.1.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pytz==2019.3
requests
requests-oauthlib==1.3.0
rsa==4.0
six==1.12.0
tzlocal==2.0.0
uritemplate==3.0.0
urllib3==1.25.6
Werkzeug==0.16.0
